carrito y gitignore

+++ carrito.js
@@ -11,15 +11,20 @@ class Carrito {
     return await this.contenedor.save({ productos: [] });
   }
 
+  // Devuelve true si se borró correctamente
   async deleteCarrito(id) {
-    await this.contenedor.deleteById(id);
+    if (await this.esIdValido(id)) {
+      await this.contenedor.deleteById(id);
+      return true;
+    } else {
+      return false;
+    }
   }
 
   // Devuelve null en caso de no encontrar el carrito
-
   async getProductos(id) {
     const carrito = await this.contenedor.getById(id);
-    if (carrito !== null) {
+    if (carrito != null) {
       return carrito.productos;
     } else {
       return null;
@@ -28,9 +33,9 @@ class Carrito {
 
   async postProducto(idCarrito, idProducto) {
     const producto = await this.productos.getProductoById(idProducto);
-    if (producto !== null) {
+    if (producto != null) {
       const viejoCarrito = await this.contenedor.getById(idCarrito);
-      if (viejoCarrito !== null) {
+      if (viejoCarrito != null) {
         await this.contenedor.updateById(
           { productos: [...viejoCarrito.productos, producto] },
           idCarrito
@@ -43,18 +48,31 @@ class Carrito {
 
   async deleteProducto(idCarrito, idProducto) {
     const viejoCarrito = await this.contenedor.getById(idCarrito);
-    if (viejoCarrito !== null) {
-      const productosAntes = viejoCarrito.productos;
-      const productosDespues = productosAntes.filter(
-        (elem) => elem.id != idProducto
-      );
-      await this.contenedor.updateById(
-        { productos: productosDespues },
-        idCarrito
-      );
-      return true;
+    if (viejoCarrito != null) {
+      const productos = viejoCarrito.productos;
+      const index = productos.map(elem => elem.id).indexOf(idProducto)
+      console.log(productos.map(elem => elem.id))
+      console.log(index)
+
+      if(index != -1) {
+        productos.splice(index, 1);
+        console.log(productos)
+        await this.contenedor.updateById(
+          { productos: productos },
+          idCarrito
+        );
+        return true
+      }
     }
     return false;
+  }
+
+  // Auxiliares
+
+  async esIdValido(id) {
+    const carritos = await this.contenedor.getAll();
+    const ids = carritos.map((elem) => elem.id);
+    return !isNaN(id) && ids.includes(id);
   }
 }
